:wrench: chore: need to check if the /proc/version exist before read it

+++ functions/ghc-pbpaste.fish
@@ -1,11 +1,13 @@
 function ghc-pbpaste
-  if grep -qiE "microsoft|wsl" /proc/version
-    set -l filepath $argv[1]
-    if test -z "$filepath"
-        echo "Usage: ghc-pbpaste <filepath>"
-        return 1
+  if test -e /proc/version
+    if grep -qiE "microsoft|wsl" /proc/version
+      set -l filepath $argv[1]
+      if test -z "$filepath"
+          echo "Usage: ghc-pbpaste <filepath>"
+          return 1
+      end
+      powershell.exe Get-Clipboard > $filepath
     end
-    powershell.exe Get-Clipboard > $filepath
   end
 end
